Rework getDriver todo and clock nowMs
