+++ zipline-loader/src/jvmMain/kotlin/app/cash/zipline/loader/ziplineCacheJvm.kt
@@ -1,31 +1,18 @@
 package app.cash.zipline.loader
 
+import com.squareup.sqldelight.db.SqlDriver
 import com.squareup.sqldelight.sqlite.driver.JdbcSqliteDriver
 import com.squareup.sqldelight.sqlite.driver.JdbcSqliteDriver.Companion.IN_MEMORY
 import okio.FileSystem
 import okio.Path
 
-actual fun openZiplineCache(
-  fileSystem: FileSystem,
-  dbPath: Path,
-  directory: Path,
-  maxSizeInBytes: Long,
-): ZiplineCache {
-  val driver = JdbcSqliteDriver(IN_MEMORY + dbPath.toString())
-  val ziplineCache = ZiplineCache(
-    driver = driver,
-    fileSystem = fileSystem,
-    directory = directory,
-    maxSizeInBytes = maxSizeInBytes,
-  )
-  ziplineCache.prune()
-  return ziplineCache
-}
+actual fun getDriver(path: Path): SqlDriver = JdbcSqliteDriver(IN_MEMORY + path.toString())
 
 fun openZiplineCacheForTesting(
   fileSystem: FileSystem,
   directory: Path,
   maxSizeInBytes: Long,
+  nowMs: () -> Long
 ): ZiplineCache {
   val driver = JdbcSqliteDriver(IN_MEMORY)
   val ziplineCache = ZiplineCache(
@@ -33,6 +20,7 @@ fun openZiplineCacheForTesting(
     fileSystem = fileSystem,
     directory = directory,
     maxSizeInBytes = maxSizeInBytes,
+    nowMs = nowMs,
   )
   ziplineCache.prune()
   return ziplineCache
